+++ angular_components/lib/material_input/material_input_multiline.dart
@@ -167,7 +167,7 @@ class MaterialMultilineInputComponent extends BaseMaterialInput
   /// Textarea element tabindex.
   ///
   /// Disabled textarea is not interactive and should not receive focus on TAB.
-  int get inputTabIndex => disabled! ? -1 : 0;
+  int get inputTabIndex => disabled ? -1 : 0;
 
   @visibleForTemplate
   void handleChange(Event event, TextAreaElement element) {