+++ angular_components/CHANGELOG.md
@@ -16,6 +16,7 @@
 * Added `ChangeDetectionStrategy.OnPush` to `material_tree_dropdown`
 * Added `ChangeDetectionStrategy.OnPush` to `material_tree_filter`
 * Added `ChangeDetectionStrategy.OnPush` to `material_tree_impl`
+* Updated `disabled` to be non nullable
 
 ## 1.0.3